requirements.txt
logipy==1.3.1
paho-mqtt==1.6.1
pyyaml==6.0.2
